pyproject.toml
[project]
name = "cnc-project"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fonttools>=4.60.0",
    "hershey-fonts>=2.1.0",
    "matplotlib>=3.10.6",
    "pyserial>=3.5",
    "setuptools>=80.9.0",
    "streamlit>=1.50.0",
    "svgpathtools>=1.7.1",
]